Fix #5 Invalid nested keywords should be ignored
If keyword nesting happens where the prior keyword is not an ONBUILD,
no hover support should be provided. However, if the parser ends up
running to the beginning of the file while looking for the initial
keyword, the parser does not consider that keyword's validity. The
parser should always check whether the previous word that has been
found is an ONBUILD or not prior to determining whether hover
information should be returned.

Signed-off-by: Remy Suen <remy.suen@gmail.com>