Fix #32 Create a new common NO_SOURCE_IMAGE error
Originally, when the first instruction was not a FROM, a
FROM_NOT_FIRST error would be created. Also, if a FROM instruction
was not found anywhere in the Dockerfile, a FROM_MISSING error would
be created.

If the first instruction is not a FROM or if the Dockerfile doesn't
have any instructions, we should just notify the user that a source
image must be specified with a FROM instruction. Having a FROM
instruction anywhere else in the file does not help if the first
instruction is not a FROM. As such, having a different error message
to differentiate the two does not help the user in any meaningful
way.

Signed-off-by: Remy Suen <remy.suen@gmail.com>