Fix #27 Correct the URI of the opened document
When a document was opened, the wrong URI was being stored by the
server. Thus, even though a validation request was processed, the
client would not show any diagnostics because the wrong URI was being
sent back to the client. Correcting the URI will now get the client
to render the appropriate diagnostics when a file is first opened
without needing it to be modified first.

Signed-off-by: Remy Suen <remy.suen@gmail.com>