Fix #62 Use the diagnostic's range for code actions
When a selection of text is made, the range that is provided for the
textDocument/codeAction request includes this range. We should not be
using that range for the generated commands but to instead use the
diagnostic's range so that we only modify the problematic text and
not all of the selected text.

Signed-off-by: Remy Suen <remy.suen@gmail.com>