Fix #64 Allow comments and blank lines after an escape
Normally, an escape character is followed by a newline and then the
rest of the instruction's arguments would follow in the next line.
However, for RUN instructions, the arguments often get very, very
long. In these cases, it may be helpful to inline comments throughout
the arguments to the instruction. We should tolerate these comments
as well as blank lines when parsing an instruction.

Signed-off-by: Remy Suen <remy.suen@gmail.com>