Correct the test setup
When testing that a \$var is correctly escaped and not defined as a
variable, we need to original variable declaration in the ARG
instruction to define a value for the variable. Otherwise, a null
will most certainly be returned regardless of whether the escape is
working or not as the variable has no value defined.

Signed-off-by: Remy Suen <remy.suen@gmail.com>