Fix #52 Ignore escape if following character is not valid
Escape characters are generally expected to be used for escaping
newlines or spaces or the $ symbol for variables. However, the user
may choose to insert it in other places as well. In the event that
the character that follows an escape is not recognized, it should be
stripped from the internal representation of the argument so that it
will not be interpreted by the validator.

Signed-off-by: Remy Suen <remy.suen@gmail.com>