Fix #65 Don't validate unrecognized parser directives
As comments can be used for anything, we shouldn't give the user an
error if what seems to be a parser directive is invalid. The user may
simply just be putting a comment at the top of the file so we
shouldn't try to validate something unless we are sure that it is a
parser directive.

Signed-off-by: Remy Suen <remy.suen@gmail.com>