WIP #26 Use new parser to check if indentation is needed
Instead of iterating over the document directly, we can use the new
parser to check if the selected lines are contained within an
instruction. If the selected lines are contained within an
instruction and it is not the first line of the instruction, then we
know that indentation will at least be required on the first line of
the selection.

Signed-off-by: Remy Suen <remy.suen@gmail.com>