Fix #70 Format nested empty lines and comments
When an instruction, like a RUN, spans multiple lines, it is not
uncommon for there to be empty newlines consisting only of whitespace
or comments to be one of those many lines for clarity purposes. As
these lines are considered to be part of the instruction from the
parser's point of view, these lines should also be formatted when a
formatting request has been received from the client.

Signed-off-by: Remy Suen <remy.suen@gmail.com>