WIP #68 First pass at supporting ENV variables
Refactored the ARG argument handling code and then replicated it in
ENV. ENV variables that are defined in the same way via the use of a
equals symbol delimiter now works. However, only one variable is
supported at the moment even though the equals symbol syntax on an
ENV means that multiple environment variables may be defined with one
instruction.

Signed-off-by: Remy Suen <remy.suen@gmail.com>