Fix #81 Warn if multiple CMD instructions are found
While it is possible for a Dockerfile to contain multiple CMD
instructions, the last one found in the file is the only one that
will be used. If there are multiple CMD instructions in a Dockerfile,
the server will now optionlly create a diagnostic to notify the user
about this.

Signed-off-by: Remy Suen <remy.suen@gmail.com>