Fix #85 Warn if HEALTHCHECK's retries flag is NaN
As the number of retries for a HEALTHCHECK should be an integer, the
validator should check and handle this case and create a diagnostic
if the value cannot be parsed into an integer. Note that while
decimal numbers such as 3.0 can be parsed into an integer, the Docker
builder will throw an error so this case must be considered as well.

Signed-off-by: Remy Suen <remy.suen@gmail.com>