Fix #89 Warn if --retries of HEALTHCHECK is not positive
The retries flag of a HEALTHCHECK specifies the number of times that
the HEALTHCHECK CMD should be run before the container is considered
to be `unhealthy`. As such, the retries flag should be a positive
integer as negative values or "zero retries" don't make any sense. A
diagnostic will now be created to handle this error in Dockerfiles.

Signed-off-by: Remy Suen <remy.suen@gmail.com>