Fix #90 Warn if multiple ENTRYPOINT instructions found
A Dockerfile may have any number of ENTRYPOINT instructions. However,
even if multiple are defined, the last one is the only one that will
take effect. The server will now check the number of ENTRYPOINT
instructions in the file and if more than one are found, a diagnostic
will be created for this problem. Its severity may be toggled and it
is also possible to completely ignore this problem.

Signed-off-by: Remy Suen <remy.suen@gmail.com>