Fix #95 Prevent hovers from showing for invalid flags
The hovering code uses the relevant word at the cursor's location to
if it's a HEALTHCHECK flag. However, if the word itself is correct
but the context of it is not then the word still gets displayed. It
should instead consider the location of the word relative to the
context of its position in the HEALTHCHECK instruction to determine
whether whether any content should be displayed or not.

Signed-off-by: Remy Suen <remy.suen@gmail.com>