Fix #99 Convert diagnostics' codes to a number
According to the protocol, a Diagnostic's code may be a number or a
string. Our current assumption that it is always a number will cause
code actions to not be calculated properly when analyzing the
provided diagnostic from the client. Converting the `code` property
if it is a string to a number will allow the server to return code
actions properly.

Signed-off-by: Remy Suen <remy.suen@gmail.com>