Fix #107 Fix parsing of duration flags
We currently assume that a duration value of zero means that a
duration has not been specified. This is incorrect because a duration
of zero may have been specified. Instead, we should mark when a time
unit has been specified and then just use that to determine whether
the flag has time units specified or not.

Signed-off-by: Remy Suen <remy.suen@gmail.com>