Fix #106 Do not suggest start-period flag with zero
As HEALTHCHECK's start-period flag cannot have a value that's less
than one millisecond, having its completion item define the value to
zero seconds is misleading as it immediately leads to a validation
error. Changed the value to be five seconds to prevent a validation
error from occurring after inserting the text of the completion item.

Signed-off-by: Remy Suen <remy.suen@gmail.com>