Fix #109 Consider position of CMD or NONE for suggestions
Instead of simply showing no suggestions if a CMD or a NONE is found
for an ONBUILD HEALTHCHECK trigger, consider where that argument is
relative to the text cursor. If it comes before the text cursor then
no flags should be suggested. If the text cursor comes before, then
flags should be suggested as completion items.

Signed-off-by: Remy Suen <remy.suen@gmail.com>