Fix #125 Correct analysis of a position in a range
The code for determining whether a given position was contained
within a range was incorrectly assuming that the range would start and
end on the same line. If the range spanned multiple lines, the
analysis would be incorrect as it needed the given position to be
contained within the starting and ending characters of the range
relative to their lines. If the range started at line 2 character 0
and ended at line 3 character 4, then any position on line 2 past
character 4 would not past the check as it would not pass the check
of being 0 <= position.character <= 4.

The fix is to consider every case individually should the position's
line be equal to the starting line, be equal to the ending line, or
is contained within somewhere between the two lines should the
range's starting and ending lines not be equal.

Signed-off-by: Remy Suen <remy.suen@gmail.com>