WIP #134 Use actual file for docker-langserver binary
It seems that referencing the server.js directly in package.json does
not appear to pass on the arguments on the command line to the
script. As other Node.js projects that uses a binary seems to
explicitly create a separate script for the launching of an
executable, we have done the same and created a bin/docker-langserver
file to startup lib/server.js.

Signed-off-by: Remy Suen <remy.suen@gmail.com>