Fix #136 Remove item arguments if snippets are disabled
The label of completion items always include its arguments. However,
if snippets are disabled, the arguments will not be appended to the
document. This can be confusing to the user because the expectation
may be that the arguments would be appended. The arguments have been
removed from the completion items if snippets are not enabled to not
mislead the user.

Signed-off-by: Remy Suen <remy.suen@gmail.com>