Fix #140 Always go to the first variable declaration
For textDocument/definition requests, the server always returned that
the definition of a variable was the last ARG instruction that
referenced that variable because of the way the code traversed the
Dockerfile's instructions. The code has been rewritten to first
identify the variable that the text cursor is on and to then traverse
the Dockerfile from top to bottom and to return the first ARG or ENV
instruction that declares the variable.

Signed-off-by: Remy Suen <remy.suen@gmail.com>