Fix #154 Add completion items for image tags in FROM
When entering in the name of an image, tags are often used to ensure
that the image is stable since the 'latest' tag will be replaced as
new images are pushed to the Docker registry. To make it easier for
users to enter in a tag of an image, we now query the official public
Docker registry using the Docker Registry HTTP API V2 for the names
of the specified image and display them to the user.

Signed-off-by: Remy Suen <remy.suen@gmail.com>