Fix #156 Do not suggest duplicated build stage names
If the same build stage name is used multiple times in the
Dockerfile, only one instance of it will be suggested as a completion
item for COPY's --from flag. As the casing of a build stage name does
not matter, only the one that was declared earlier in the file will
be suggested instead of having multiples show up in the list.

Signed-off-by: Remy Suen <remy.suen@gmail.com>