Fix #159 Set documentation for suggested build stages
When a build stage is suggested, it may not be completely obvious
what the source image of that build stage is. To alleviate this, we
now display set the source image of the build stage as the
documentation text of the provided completion item. This should help
users easily figure out what the source image of a suggested build
stage is.

Signed-off-by: Remy Suen <remy.suen@gmail.com>