Fix #164 Do not query the registry for image digests
Due to the format of an image digest to be used in a FROM
instruction, it is currently possible for the parser to incorrectly
think that the FROM instruction is pointing at an image's tag instead
of an image's digest. The parser has been fixed to correctly
recognize whether a FROM is pointing at an image's tag or an image's
digest.

Signed-off-by: Remy Suen <remy.suen@gmail.com>