Fix #153 Scope FROM variables correctly
Variables that are used in a FROM instruction can only interact with
the initial set of ARG instructions at the top before the first FROM
instruction. ENV variables that are declared at the top of a
Dockerfile cannot be processed by the Docker builder and should not
be connected with any FROM instruction that refers to such a variable
when processing a textDocument/definition, textDocument/hover,
textDocument/documentHighlight, or textDocument/rename request.

Signed-off-by: Remy Suen <remy.suen@gmail.com>