WIP #162 Fix argument handling of ENV and LABEL
When multiple whitespace characters are used to delimit multiple ENV
variables or LABEL metadata properties, the parser will incorrectly
identify extra arguments in the instruction. This causes an issue
when trying to process the signature information at the current text
cursor because the extra argument's range will conflict with the
expected signature parameter at the current position. The parser will
now ignore multiple whitespace characters when trying to process the
arguments of an ENV or LABEL instruction and not automatically assume
that there is a new argument to process.

Signed-off-by: Remy Suen <remy.suen@gmail.com>