Fix #168 Ignored nested comments in ARG and ENV
If an ARG or ENV instruction is using a single or double quote to
define a literal value for a variable, nested comments in an escaped
newline are being incorrectly included in the variable's value. The
parser will now ignore such commen when calculating the value of a
variable for hovers.

Signed-off-by: Remy Suen <remy.suen@gmail.com>