Fix #179 Implement support for workspace/configuration
The new workspace/configuration request that's being proposed will
allow the validator's settings to not be global but to instead be
scoped to a given resource. This means that the generated diagnostics
of one Dockerfile can differ from that of another Dockerfile even
when they are both communicating with the same language server
instance.

Signed-off-by: Remy Suen <remy.suen@gmail.com>