Fix #177 Warn about empty continuation lines
If an instruction spans multiple lines but an escaped newline is
followed by an empty line that consists only of whitespace characters
instead of content, a warning will be flagged by the Docker CE 17.09
builder. As it is a warning and not an error, the language server
will default to publishing a warning diagnostic for this case but the
severity of the generated diagnostic can be toggled via the new
docker.languageserver.diagnostics.emptyContinuationLine setting.

Signed-off-by: Remy Suen <remy.suen@gmail.com>