Fix #184 Suggest variables even if ARG is malformed
In order to suggest variables to the user for a
textDocument/completion request, we must examine all the ARG and ENV
instructions that exist prior to the current cursor location to find
all the applicable variable names that should be suggested. If an ARG
instruction has no variable defined, the code should skip over that
ARG and not bother trying to read its non-existent variable's name to
suggest to the user.

Signed-off-by: Remy Suen <remy.suen@gmail.com>