Fix #183 Check before sending workspace/applyEdit
The language server currently provides code actions that point to
commands that are handled by the server via the use of the
workspace/applyEdit request. However, clients can opt to not support
the workspace/applyEdit request. In such a scenario, the server
should not provide any code actions because they can't be handled by
the server. The server should also not honour any
workspace/executeCommand requests as they cannot be fulfilled if the
client does not support workspace/applyEdit.

Signed-off-by: Remy Suen <remy.suen@gmail.com>