Fix #178 Fix broken socket support
Changed the dependency to vscode-languageserver to fix the module's
broken socket support. The client should listen for connections on a
port and then pass that port in when starting up the language server.
The language server will then connect to that port and the client and
server can then start communicating with each other via the socket.

Signed-off-by: Remy Suen <remy.suen@gmail.com>