Fix #186 Correct range for unknown flag diagnostics
If the validator only uses the flag's name's range for the published
diagnostic, the range may potentially be of length zero if the flag
doesn't have a name if it simply the string "--". The handling and
rendering of diagnostics with zero length ranges is ambiguous and may
differ between clients so the diagnostic has been changed to instead
include the "--" prefix in its range so that its length is always at
least two.

Signed-off-by: Remy Suen <remy.suen@gmail.com>