Fix #79 Add CLI frontend for formatting and validating
While many of the features provided by the language server protocol
only make sense in the context of an editor, formatting and
validation are both operations that can be performed on a Dockerfile
even if they haven't been opened in an editor. These two features
have been added to the docker-langserver binary so that a user can
now format and validate Dockerfiles straight from the command line
without the use of any editors.

Signed-off-by: Remy Suen <remy.suen@gmail.com>