Fix #190 Don't show completion items in comments
The ranges of comments were being set incorrectly by the parser which
caused the positional analysis to think that the user was not inside
a comment when the user actually was. This caused completion items to
be displayed to the user even when the user was typing inside a
comment.

Signed-off-by: Remy Suen <remy.suen@gmail.com>