Fix #191 Single quote LABEL values should be literal
If an environment variable such as a $var or ${var} is found in a
LABEL's value that is contained in single quotes, they should be
ignored. This is because the content in a single quoted string should
be read literally and not go through any processing for variable
expansions and such.

Signed-off-by: Remy Suen <remy.suen@gmail.com>