Fix #200 Merge default variables so they only show once
When a textDocument/completion request is received for variables, the
predefined ARG variables that Docker provides should be suggested.
However, if a variable with the same name has been defined in the
Dockerfile, that variable should not show up as two different
completion items. Instead, they should be merged so that only one
shows up.

Signed-off-by: Remy Suen <remy.suen@gmail.com>