Fix #205 Replaced existing code with a new npm module
The code for the analysis and processing of a Dockerfile for the
returning of information has been refactored out into a new
dockerfile-language-service module.

This module now only contains the necessary code for the language
server process itself. The module will forward the requests from the
client to the language service module and also forward the responses
from the language service module back to the client.

Signed-off-by: Remy Suen <remy.suen@gmail.com>