Fix #202 Use versioned edits if client supports it
An LSP client can declare that it supports versioned edits for a
workspace/applyEdit request via the documentChanges client
capability. If this is enabled, the server should send back text
edits that are versioned so that they will only be applied to the
document in the client if its version matches that of the sent text
edits.

Signed-off-by: Remy Suen <remy.suen@gmail.com>