Fix #216 Apply textDocument/didChange changes in order
When changes are received from a textDocument/didChange notification,
each change represents a state change of the document from one state
to the next with each change relative to the state after the previous
change has been applied. Thus, it is actually incorrect and
unnecessary to sort them and to then apply them backwards. The
changes should simply be applied in the order in which they are
defined in the JSON array.

Signed-off-by: Remy Suen <remy.suen@gmail.com>