Fix #218 Consolidate how settings are read
A bad case of copy/paste caused the wrong setting to be read by the
validator. instructionJSONInSingleQuotes was also not read if a
client did not support workspace/configuration requests and instead
used a global setting.

The code for reading the settings has been consolidated to hopefully
prevent this mismatch from happening again in the future.

Signed-off-by: Remy Suen <remy.suen@gmail.com>